+++ src/markdown/index.test.ts
@@ -1143,7 +1143,7 @@ describe("test generateMarkdown", () => {
   });
 
   describe("test validation", () => {
-    it("should end process on wrong example provided", () => {
+    it("should throw on wrong example provided", () => {
       const testSchema: SpecJsonSchemaRoot = {
         $id: "http://example.com/schemas/test-schema",
         title: "Test Schema",
@@ -1168,15 +1168,15 @@ describe("test generateMarkdown", () => {
       };
 
       const spyOnLogError = mock.spyOn(log, "error");
-      const spyOnProcessExit = mock.spyOn(process, "exit").mockImplementation(() => undefined as never);
 
-      generateMarkdown(testSchema, specId, "spec", undefined);
+      expect(() => generateMarkdown(testSchema, specId, "spec", undefined)).toThrow(
+        'Example value "12345" is invalid: must be string',
+      );
 
       expect(spyOnLogError).toHaveBeenCalledWith(expect.stringContaining(`Example value "12345" is invalid:`));
-      expect(spyOnProcessExit).toHaveBeenCalledWith(1);
     });
 
-    it("should end process on wrong default value provided", () => {
+    it("should throw on wrong default value provided", () => {
       const testSchema: SpecJsonSchemaRoot = {
         $id: "http://example.com/schemas/test-schema",
         title: "Test Schema",
@@ -1199,12 +1199,12 @@ describe("test generateMarkdown", () => {
       };
 
       const spyOnLogError = mock.spyOn(log, "error");
-      const spyOnProcessExit = mock.spyOn(process, "exit").mockImplementation(() => undefined as never);
 
-      generateMarkdown(testSchema, specId, "spec", undefined);
+      expect(() => generateMarkdown(testSchema, specId, "spec", undefined)).toThrow(
+        'Default value "12345" is invalid: must be string',
+      );
 
       expect(spyOnLogError).toHaveBeenCalledWith(expect.stringContaining(`Default value "12345" is invalid:`));
-      expect(spyOnProcessExit).toHaveBeenCalledWith(1);
     });
   });
 });